all working

+++ ajenti1/recipes/default.rb
@@ -6,17 +6,10 @@
 #
 #
 
-#include_recipe "yum"
-
-#execute "update" do
-#	command "yum update -y"
-#	action :run
-#end
-
-
 execute "enable-epel-repository" do
 	command "yum-config-manager --quiet --enable epel"
 	action :run
+	not_if "yum repolist | grep epel"
 end
 
 execute "update" do
@@ -45,24 +38,25 @@ end
 
 package ["ajenti","ajenti-v","ajenti-v-mail","ajenti-v-nginx","ajenti-v-mysql","ajenti-v-php7.0-fpm","ajenti-v-php-fpm","php70-mysqlnd","php70-fpm"]
 
-if `ls -l /etc/alternatives/php | grep 7`.empty?
-	execute 'configure php7' do
+execute 'configure php7' do
 	command "unlink /etc/alternatives/php && unlink /etc/alternatives/php-fpm && unlink /etc/alternatives/php-fpm-init && ln -s /usr/bin/php7 /etc/alternatives/php && ln -s /usr/sbin/php-fpm-7.0 /etc/alternatives/php-fpm && ln -s /etc/rc.d/init.d/php-fpm-7.0 /etc/alternatives/php-fpm-init"
 	action :run
-	end
+	not_if "ls -l /etc/alternatives/php | grep 7"
 end
 
 service 'ajenti' do
     action [:enable, :start]
 end
 
-if `cat /etc/passwd | grep www-data`.empty?
-	execute 'apply ajenti-v' do
+service 'ajenti' do
+	action [:restart]
+	not_if "cat /etc/passwd | grep www-data"
+end
+execute 'apply ajenti-v' do
 	command 'ajenti-ipc v apply'
 	action :run
-	end
+	not_if "cat /etc/passwd | grep www-data"
 end
-# Will fail here since directory probably isn't created yet at /srv/$DOMAIN
 
 execute 'import first website' do
 	command 'ajenti-ipc v import /home/ec2-user/rt/website.json && rm /home/ec2-user/rt/website.json && ajenti-ipc v apply'
